+++ src/main/java/com/pingplusplus/model/RoyaltySettlement.java
@@ -11,10 +11,10 @@ public class RoyaltySettlement extends AccountAPIResource {
     String object;
     Boolean livemode;
     Long created;
-    Integer amount;
-    Integer amountSucceeded;
-    Integer amountFailed;
-    Integer amountCanceled;
+    Long amount;
+    Long amountSucceeded;
+    Long amountFailed;
+    Long amountCanceled;
     Integer count;
     Integer countSucceeded;
     Integer countFailed;
@@ -60,35 +60,35 @@ public class RoyaltySettlement extends AccountAPIResource {
         this.created = created;
     }
 
-    public Integer getAmount() {
+    public Long getAmount() {
         return amount;
     }
 
-    public void setAmount(Integer amount) {
+    public void setAmount(Long amount) {
         this.amount = amount;
     }
 
-    public Integer getAmountSucceeded() {
+    public Long getAmountSucceeded() {
         return amountSucceeded;
     }
 
-    public void setAmountSucceeded(Integer amountSucceeded) {
+    public void setAmountSucceeded(Long amountSucceeded) {
         this.amountSucceeded = amountSucceeded;
     }
 
-    public Integer getAmountFailed() {
+    public Long getAmountFailed() {
         return amountFailed;
     }
 
-    public void setAmountFailed(Integer amountFailed) {
+    public void setAmountFailed(Long amountFailed) {
         this.amountFailed = amountFailed;
     }
 
-    public Integer getAmountCanceled() {
+    public Long getAmountCanceled() {
         return amountCanceled;
     }
 
-    public void setAmountCanceled(Integer amountCanceled) {
+    public void setAmountCanceled(Long amountCanceled) {
         this.amountCanceled = amountCanceled;
     }
